fetch-xml.sh: improve metadata refresh timing workaround
The workaround for the timing issue was still creating a gap: the timestamp was
updated after fetching the file, but there was a gap while the file was being
validated.

Improve this workaround to set the timestamp AFTER moving the file into the
target location.

See https://reannz.atlassian.net/browse/TUAK-70 for internal details and
https://issues.shibboleth.net/jira/browse/OSJ-165 for upstream issue (tracked
against OpenSAML3)